[PATCH 124/194] power: supply: axp20x_usb_power: allow disabling input current limiting
The AXP PMICs allow the user to disable current limiting on the VBUS
input. While read-out of this setting was already supported by the
driver, it did not allow the user to configure the PMIC to disable
current limiting.

Add support for this.

Signed-off-by: Chen-Yu Tsai <wens@csie.org>
Acked-by: Maxime Ripard <maxime.ripard@bootlin.com>

Gbp-Pq: Topic power
Gbp-Pq: Name 0038-power-supply-axp20x_usb_power-allow-disabling-input-.patch