[PATCH 005/194] bus: sunxi-rsb: Precompute read/write commands and mask
Since we know the size of the transfer at context creation time, go
ahead and determine the commands used for reading and writing then,
instead of running through a switch statement for every read/write.
We can do the same thing for the read mask.

The context pointer could be passed directly to sunxi_rsb_read/write
to avoid adding more parameters (which would be spilled for mutex_lock).
That would make the regmap read/write wrappers even more trivial, so I
inlined sunxi_rsb_read/write into them.

I changed the error message to print the name of the device requesting
the regmap, not the RSB controller; I expect that to be more helpful
when tracking down the source of the error.

Signed-off-by: Samuel Holland <samuel@sholland.org>

Gbp-Pq: Topic common
Gbp-Pq: Name 0005-bus-sunxi-rsb-Precompute-read-write-commands-and-mas.patch