[PATCH 003/194] bus: sunxi-rsb: Always check register address validity
The register address was already validated for read operations before
being truncated to a u8. Write operations have the same set of possible
addresses, and the address is being truncated from u32 to u8 here as
well, so the same check is needed.

Signed-off-by: Samuel Holland <samuel@sholland.org>

Gbp-Pq: Topic common
Gbp-Pq: Name 0003-bus-sunxi-rsb-Always-check-register-address-validity.patch