arm: dts: pinephone: change brightness scale
A brightness level < 21 totally disables backlight, so starting with a
level of 30 makes sense, as disabling the backlight can be achieved by
pressing the power button.

Additionally, gnome-settings-daemon considers 0 an acceptable value if
there are less than 99 different brightness levels, which disables the
backlight too. Therefore we set 100 brightness levels, so that the user
can set the brightness to its minimum value without inadvertently
disabling the backlight (and then struggle to get his display back).