[PATCH 062/194] ASoC: sun50i-codec-analog: Add support for internal bias
In order to properly bias headset microphones, there should be a
resistor between pins HBIAS and MICDET. This can be an external
resistor, but the codec also provides an internal 2.2K resistor which
can be enabled by a register.

This patch enables or disables the internal bias resistor based on the
appropriate device tree property.

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>
[Samuel: split binding and implementation patches]
Signed-off-by: Samuel Holland <samuel@sholland.org>

Gbp-Pq: Topic audio
Gbp-Pq: Name 0079-ASoC-sun50i-codec-analog-Add-support-for-internal-bi.patch