sun8i-codec: add binding for inverted jack detection
Depending on how the jack connector is wired on the board, jack
detection IRQ's can be inverted, i.e the IN IRQ can be triggered when
unplugging the jack while the OUT IRQ will be triggered when a jack is
inserted.

This commit introduces a DT property so the driver can be made aware of
such a case and process the IRQ's accordingly.

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>