sun8i-codec: enable jack detection externally
As jack detection isn't supported by all variants of this codec and has
to be enabled for the analog codec too, it doesn't make sense to enable
it here by default.

This commit introduces an exported function for the analog codec to
enable jack detection if needed. The analog codec will be responsible
for providing a callback to enable button detection on its side.

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>