[PATCH 184/194] arm64: dts: allwinner: pinephone: improve device tree
On PinePhone, the headset mic bias resistor isn't populated on the
schematics (R811), therefore we need to enable the codec's internal
resistor. Additionnally, the jack detection IRQ's are inverted due to the
connector wiring, so the necessary property is added to the codec node
to made the driver aware of this fact.

We also stop LEDs during suspend to improve battery life, lower
cpu_alert* temperatures so the phone doesn't get too hot and improve the
backlight brightness values so we have a wider usable range.

Finally, the RGB LED max_brightness is set to 1 as it isn't using a PWM
output.

Gbp-Pq: Topic pinephone
Gbp-Pq: Name 0170-arm64-dts-allwinner-pinephone-improve-device-tree.patch