arm64: dts: allwinner: pinephone: enable jack detection
On PinePhone, the headset mic bias resistor isn't populated on the
schematics (R811), therefore we need to enable the codec's internal
resistor.

Additionnally, the jack detection IRQ's are inverted due to the
connector wiring, so the necessary property is added to the codec node
to made the driver aware of this fact.

Finally, add the "PinePhone" name to the sound card: this will make
upstreaming ucm config easier as we can use a unique name.

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>