[PATCH 127/194] regulator: axp20x: Turn N_VBUSEN to input on x-powers,sense-vbus-en
When x-powers,sense-vbus-en is present on the PMIC DT node, force
the N_VBUSEN to be an input. This is necessary on boards that
have N_VBUSEN connected to VBUS enabling switch to make PMIC
not charge when the board is generating VBUS voltage itself.

Signed-off-by: Ondrej Jirman <megous@megous.com>

Gbp-Pq: Topic pinephone
Gbp-Pq: Name 0163-regulator-axp20x-Turn-N_VBUSEN-to-input-on-x-powers-.patch