[PATCH 041/194] drm/panel: st7703: Make the sleep exit timing match the spec
The driver seemed to try to make the total cumulative time of all delays
across the whole power up sequence to sum up to 120ms.

The datasheet on the other hand specifies that there needs to be 120ms
delay after the sleep out command, not after reset as the driver
assumes.

Lastly, the delay between init commands and the sleep exit command is
not necessary. (not specified anywhere)

Reported-by: Samuel Holland <samuel@sholland.org>
Signed-off-by: Ondrej Jirman <megous@megous.com>

Gbp-Pq: Topic display
Gbp-Pq: Name 0047-drm-panel-st7703-Make-the-sleep-exit-timing-match-th.patch