[PATCH 061/194] ASoC: dt-bindings: sun50i-codec: Add binding for internal bias
In order to properly bias headset microphones, there should be a
resistor between pins HBIAS and MICDET. This can be an external
resistor, but the codec also provides an internal 2.2K resistor which
can be enabled by a register.

This patch adds a device-tree property to the sun50i-codec-analog driver
to take advantage of this feature.

Signed-off-by: Arnaud Ferraris <arnaud.ferraris@collabora.com>
[Samuel: split binding and implementation patches]
Signed-off-by: Samuel Holland <samuel@sholland.org>

Gbp-Pq: Topic audio
Gbp-Pq: Name 0078-ASoC-dt-bindings-sun50i-codec-Add-binding-for-intern.patch