usb: typec: anx7688: Update input current limit respecting PD SRC_CAPs
Previously the limit was bumped solely based on CC pins, which is not
standards compliant. The device should wait for PD power negotiation
before falling back to using resistors on CC pins to see what it can
sink.

Let's still use auto_pd, but wait for SRC_CAPs. FW will send SRC_CAPs
when PD it generates the RDO for the SOURCE, but before source ACCEPTS
it.

We perform input current update with a timer:

- 3s after connecting the cable if no PD source caps are received
- 0.5s after PD source caps are received

It's still possible that RDO generated by FW will get rejected,
but unlikely. In that situation we would overload the source. That
usually leads to hard reset and restart of the negotiation. This
may lead to infinite loop. I guess it's unlikely.

It's probably better to wait for ACCEPT on the RDO. We can't recover
anyway while using auto_pd mode of the FW.

Signed-off-by: Ondrej Jirman <megous@megous.com>

Gbp-Pq: Topic anx7688
Gbp-Pq: Name 0123-usb-typec-anx7688-Update-input-current-limit-respect.patch