anx7688: Don't reset input current limit to 500mA on probe
This cancels the result of the current result of BC1.2 detection,
and doesn't help with anything really.

With this patch, if the phone was plugged to a dumb charger when
off, and then booted, the initial BC1.2 detection result done during the
charger plugin will be kept.

Also fix race between interrupt and "enabling USB BC 1.2 detection\n"

Signed-off-by: Ondrej Jirman <megous@megous.com>

Gbp-Pq: Topic anx7688
Gbp-Pq: Name 0129-anx7688-Don-t-reset-input-current-limit-to-500mA-on-.patch