usb: typec: anx7688: Drop use of nopd_timer, use a ktime based deadline
We already have one periodic timer that re-checks status of the anx7688
and can be used to finish the update of the current allowance after
PD/BC1.2 negotiation settles.

Simplify the code by just setting a deadline when the PD should have
already happened, and check that deadline whenever we call our work
function. This makes it easier to do the right thing across suspend/resume
cycle.

Signed-off-by: Ondrej Jirman <megous@megous.com>

Gbp-Pq: Topic anx7688
Gbp-Pq: Name 0130-usb-typec-anx7688-Drop-use-of-nopd_timer-use-a-ktime.patch